requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
pandas==2.2.2
openpyxl==3.1.5
python-dateutil==2.9.0.post0
